fix: possible fix for the corrupted taskbar "file" issue by making sure it makes a FOLDER not a FILE

+++ src/playbook/Executables/TASKBARPINS.ps1
@@ -84,6 +84,9 @@ foreach ($entry in $regTaskbar.GetEnumerator()) {
 $taskBarLocation = 'Microsoft\Internet Explorer\Quick Launch\User Pinned\TaskBar'
 $rootKey = 'SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Taskband'
 
+Stop-Process -Name explorer -Force -ErrorAction SilentlyContinue
+Start-Sleep -Milliseconds 500
+
 # Clearing taskbar, copying the shortcut, setting registry
 foreach ($userKey in (Get-RegUserPaths -NoDefault).PsPath) {
     $sid = Split-Path $userKey -Leaf
@@ -93,17 +96,22 @@ foreach ($userKey in (Get-RegUserPaths -NoDefault).PsPath) {
         Write-Error "Couldn't find AppData value for $sid!"
     } else {
         Write-Title "Setting '$Browser' taskbar shortcut for '$sid'..."
-        
         Write-Output "Clearing current shortcuts..."
         $taskBarAppData = "$appData\$taskBarLocation"
         if (Test-Path $taskBarAppData -PathType Leaf) {
-            Write-Output "Deleting TaskBar file..."
+            Write-Output "Deleting corrupted TaskBar file..."
             Remove-Item -Path $taskBarAppData -Force
         }
-        Get-ChildItem $taskBarAppData | Remove-Item -Force -Recurse
+        if (!(Test-Path $taskBarAppData -PathType Container)) {
+            Write-Output "Creating TaskBar folder..."
+            New-Item -Path $taskBarAppData -ItemType Directory -Force | Out-Null
+        } else {
+            Get-ChildItem $taskBarAppData | Remove-Item -Force -Recurse
+        }
 
         Write-Output "Adding new shortcuts..."
-        Copy-Item -Path "$tmp\*" -Destination $taskBarAppData -Force
+        # make sure its a folder with a backslash
+        Copy-Item -Path "$tmp\*" -Destination "$taskBarAppData\" -Force
 
         Write-Output "Changing in Registry..."
         $key = "$(Convert-Path $userKey)\$rootKey"
@@ -112,4 +120,5 @@ foreach ($userKey in (Get-RegUserPaths -NoDefault).PsPath) {
         }
     }
 }
-Stop-Process -Name explorer -Force
+
+Stop-Process -Name explorer -Force -ErrorAction SilentlyContinue
